django-server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["byamb4 <byamba4life@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-cors-headers = "^3.7.0"
Django = "^3.1.6"
djangorestframework-simplejwt = "^4.6.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
psycopg2 = "^2.8.6"
Pillow = "^8.1.0"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
